refactor: UserService interface 생성, 다형성과 OCP
