+++ apptive/src/main/java/com/backend/apptive/service/UserService.java
@@ -1,56 +1,17 @@
 package com.backend.apptive.service;
 
+import com.backend.apptive.domain.User;
 import com.backend.apptive.dto.UserDto;
 import com.backend.apptive.exception.DuplicateEmailException;
-import com.backend.apptive.repository.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
-import lombok.RequiredArgsConstructor;
-import com.backend.apptive.domain.User;
-import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-@RequiredArgsConstructor
-@Service
-@Transactional(readOnly = true)
-public class UserService {
-
-    private final UserRepository userRepository;
-
-    @Transactional
-    public void save(UserDto.Request request){
-        if (userRepository.existsByEmail(request.getEmail())) {
-            throw new DuplicateEmailException("이미 존재하는 이메일입니다: " + request.getEmail());
-        }
-        userRepository.save(request.toEntity());
-    }
-
-    public List<UserDto.Response> findAll() {
-        return userRepository.findAll()
-                .stream()
-                .map(UserDto.Response::toDto)
-                .toList();
-    }
-
-    public UserDto.Response findByEmail(String email){
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new EntityNotFoundException("유저를 찾을 수 없습니다: " + email));
-
-        return UserDto.Response.toDto(user);
-    }
-
-    @Transactional
-    public void deleteByEmail(String email){
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new EntityNotFoundException("유저를 찾을 수 없습니다: " + email));
-        userRepository.delete(user);
-    }
-
-    @Transactional
-    public void update(String email, UserDto.Request request){
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new EntityNotFoundException("유저를 찾을 수 없습니다: " + email));
-        user.update(request.getName());
-    }
+public interface UserService {
+    void save(UserDto.Request request);
+    List<UserDto.Response> findAll();
+    public UserDto.Response findByEmail(String email);
+    void deleteByEmail(String email);
+    void update(String email, UserDto.Request request);
 }
